finally rendered on the mobile screen
