+++ package.json
@@ -8,7 +8,7 @@
     "eject": "expo eject"
   },
   "dependencies": {
-    "@react-native-community/masked-view": "^0.1.11",
+    "@react-native-community/masked-view": "^0.1.10",
     "@react-navigation/bottom-tabs": "^5.11.11",
     "@react-navigation/native": "^5.9.4",
     "expo": "~41.0.1",
@@ -18,10 +18,10 @@
     "react-native": "https://github.com/expo/react-native/archive/sdk-41.0.0.tar.gz",
     "react-native-gesture-handler": "^1.10.3",
     "react-native-maps": "^0.28.0",
-    "react-native-reanimated": "^2.2.0",
+    "react-native-reanimated": "^2.1.0",
     "react-native-safe-area-context": "^3.2.0",
-    "react-native-screens": "^3.3.0",
-    "react-native-web": "~0.13.12",
+    "react-native-screens": "^3.0.0",
+    "react-native-web": "^0.11.7",
     "react-router-dom": "^5.2.0"
   },
   "devDependencies": {
